fix(gpu): remove obsolete patch from onevpl-intel-gpu
The applied patch has already been imported upstream in version 22.3.
So, it can not be applied again and build will fail.

Looking at the history of the package `onevpl-intel-gpu` in
"meta-intel" layer in yocto repository, it is totally unclear, why
this patch has ended up here. It is not needed!
Version 22.1 from Feb 2nd 2022 does not include the bad commit, that
needs to be reverted.
Version 22.3.2 introduced on March 3rd, 2022 already includes the
revert commit. Therefore, no version in Yocto contains the bad
commit and thus it seems no version check need to be performed
with the `.bbappend` file. Hence, the patch is simply removed.

see
----
Yocto Kirstone history:
https://git.yoctoproject.org/meta-intel/log/recipes-multimedia/onevpl?h=kirkstone

upstream import of removed patch
https://github.com/oneapi-src/oneVPL-intel-gpu/commit/5650a118aceb0fc098bd503e3ab5969346477a8c

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>